core/requirements.txt
# GENESIS-SOVEREIGN Requirements
# Aria's execution loop dependencies

# Core AI
ollama>=0.1.0

# Trinity AI APIs
anthropic>=0.25.0
openai>=1.12.0
google-generativeai>=0.3.0

# Code execution and testing
docker>=7.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
autopep8>=2.0.4
pylint>=3.0.3
mypy>=1.7.1

# GitHub integration
PyGithub>=2.1.1

# AWS deployment
boto3>=1.34.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
requests>=2.31.0

# Optional (for advanced features)
# kubernetes>=28.1.0  # For Kubernetes deployment
# terraform>=1.0.0    # For infrastructure as code
